+++ src/main/kotlin/App.kt
@@ -1,0 +1,28 @@
+import react.FC
+import react.Props
+import react.dom.html.ReactHTML.h3
+
+val unwatchedVideos = listOf(
+    Video(1, "Opening Keynote", "Andrey Breslav", "https://youtu.be/PsaFVLr8t4E"),
+    Video(2, "Dissecting the stdlib", "Huyen Tue Dao", "https://youtu.be/Fzt_9I733Yg"),
+    Video(3, "Kotlin and Spring Boot", "Nicolas Frankel", "https://youtu.be/pSiZVAeReeg")
+)
+
+val watchedVideos = listOf(
+    Video(4, "Creating Internal DSLs in Kotlin", "Venkat Subramaniam", "https://youtu.be/JzTeAM8N1-o")
+)
+
+val App = FC<Props> {
+    h3 {
+        +"Videos to watch"
+    }
+    VideoList {
+        videos = unwatchedVideos
+    }
+    h3 {
+        +"Videos watched"
+    }
+    VideoList {
+        videos = watchedVideos
+    }
+}
