+++ src/main/kotlin/Main.kt
@@ -1,28 +1,11 @@
-import csstype.Position
-import csstype.px
 import kotlinx.browser.document
-import react.Fragment
 import react.create
-import react.css.css
-import react.dom.html.ReactHTML.div
-import react.dom.html.ReactHTML.h1
-import react.dom.html.ReactHTML.h3
-import react.dom.html.ReactHTML.img
-import react.dom.html.ReactHTML.p
 import react.dom.render
 
 fun main() {
-    val unwatchedVideos = listOf(
-        Video(1, "Opening Keynote", "Andrey Breslav", "https://youtu.be/PsaFVLr8t4E"),
-        Video(2, "Dissecting the stdlib", "Huyen Tue Dao", "https://youtu.be/Fzt_9I733Yg"),
-        Video(3, "Kotlin and Spring Boot", "Nicolas Frankel", "https://youtu.be/pSiZVAeReeg")
-    )
-    val watchedVideos = listOf(
-        Video(4, "Creating Internal DSLs in Kotlin", "Venkat Subramaniam", "https://youtu.be/JzTeAM8N1-o")
-    )
-
     val container = document.getElementById("root") ?: error("Couldn't find root container!")
-
+    render(App.create(), container)
+    /*
     render(Fragment.create {
         h1 {
             +"Hello, React+Kotlin/JS!"
@@ -82,4 +65,5 @@ fun main() {
             }
         }
     }, container)
+     */
 }
