+++ src/main/kotlin/VideoList.kt
@@ -1,0 +1,26 @@
+import kotlinx.browser.window
+import react.FC
+import react.Props
+import react.dom.html.ReactHTML.p
+import react.key
+import react.useState
+
+external interface VideoListProps : Props {
+    var videos: List<Video>
+}
+
+val VideoList = FC<VideoListProps> { props ->
+    var selectedVideo: Video? by useState(null)
+    for (video in props.videos) {
+        p {
+            key = video.id.toString()
+            onClick = {
+                selectedVideo = video
+            }
+            if (video == selectedVideo) {
+                +"▶ "
+            }
+            +"${video.speaker}: ${video.title}"
+        }
+    }
+}
